Custom error when target field is already linked
The plugin does not allow multi-level linking of custom fields, and
triggers a generic error if user tries to set mappings to a field that
is already linked.

A specific error message was defined for this purpose in the language
strings, but it was not being used.

The plugin now displays the custom error message; the language string
was renamed to make it clear that it's an error string.